+++ cloud/src/main/java/com/project/cloud/global/common/annotation/LoginUser.java
@@ -1,4 +1,4 @@
-package com.project.cloud.global.annotation;
+package com.project.cloud.global.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
